ultimas correcciones

+++ aplicaciones/cursos_hikvision/panel/funciones/eliminar_registro.php
@@ -1,12 +1,18 @@
 <?php
+session_start();
+$h14 = $_SESSION["ing_h14"];
+if($h14 == "autorizado")
+{
+    require "../../../../conexion.php";
 
-require "../../../../conexion.php";
+    $id = $_POST["id"];
 
-$id = $_POST["id"];
+    $consulta = "delete from cursos_hikvision.evento where id=$id";
+    mysqli_query($conexion,$consulta);
 
-$consulta = "delete from cursos_hikvision.evento where id=$id";
-mysqli_query($conexion,$consulta);
-
-echo "1";
-
-?>
+    echo "1";
+}
+else
+{
+    header("Location: ../../../index.php");
+}
